EditProfileDialog: add combobox to select custom text editor
The functionality to open a file at a certain line/column in a text editor
was added in commit 923f8d144a31be. Here some preset text editors are added
to make it easier to select your preferred text editor to use; the syntax
for each editor is hard-coded, however there is a "custom" entry in the
combobox where the user can set a different text editor (or a different
syntax).

This combobox is inspired by the "Code Navigation" menu in Heaptrack.

GUI: