FileFilterHotspot: check the mime type before opening with text editor
When activating a file filter hot spot, everything got opened with the text
editor specified in the profile settings, which is wrong e.g. with pdf or
png files. Check the mime type of the file path, based only on the extension
so as to keep things fast.

Also rename openWithSysDefaultEditor() to openWithSysDefaultApp(), which is
more accurate.

BUG: 431496