FileFilter: don't create a hotspot for '................'
Only '.' and '..' could be valid hotspots, but '..........' isn't. The
latter is used by some CLI tools to indicate progress ...etc.

To test, in any dir use `ls -a`, Ctrl + clicking '.' and '..' should
open the current dir and parent of the current dir respectively.

Type .......... at the prompt, before it would get underlined, after no
hotspot will be created for it.

BUG: 432779