Changed 'Defaults' texts on reset buttons to 'Reset'
In the UI of Konsole, when editing a keyboard setting of a profile, there is a
'Defaults' button, which resets everything back to an initial state. However,
'Reset' covers its purpose more intuitively, in line with the KDE guideline on
intuitiveness and ease of use. Also, the original can be misinterpreted as:
'Make this the new defaults'. Hence the change.