Fix view splitter actions not being updated when a view is closed
This happens because in ViewManager::sessionFinished(),
the 'view' is delayed deleted which means that it is not removed
from its containing splitter until the next event.  However,
toggleActionsBasedOnState() is called immediately which makes its
decisions based on the current splitter state with the view still
present.

In order to avoid any possible object lifetime issues caused by
changing the way the view is deleted, or trying to remove it from the
view immediately, the call of toggleActionsBasedOnState() is instead
delayed until the view has been destroyed (and therefore removed from
its splitter).