Update the profiles list menu when the default profile is changed
Now the 'File->New Tab' menu is updated when the default profile is
changed, i.e. the action font is made bold and the favourites icon
emblem is added to the profile icon.

A call to menu()->setDefaultAction() has been removed, as I don't see
what it serves exactly in the current code.

Reported by khindenburg in:
https://invent.kde.org/utilities/konsole/-/merge_requests/351