De-qwidget-ize TerminalColor and TerminalFonts
Something was swallowing mouse moves on the top left 100x30 pixels of
TerminalDisplay.  As it turns out, it was the widgets for TerminalFonts
and TerminalColor.  Neither of them really needs to be a QWidget.

TerminalColor emits the onPalette(const QPalette &) signal, has slots,
and overrides event(QEvent &), so turn it into a QObject.

TerminalFonts doesn't do any of the above, turn it into a plain C++
object.

Now that TerminalFonts is no longer a QObject, manage it with a
std::unique_ptr.

TerminalFonts::fontChange(const QFont &) was calling update(), which was
probably a leftover from when it was split from TerminalDisplay.  Get
rid of it.  Trivial testing using keyboard shortcuts to change font size
(Ctrl++, Ctrl+-, Alt+Ctrl+0), and previewing font changes on "Edit
Current Profile..." seem to indicate that it isn't needed.