Restore AutoScrollHandler behavior
AutoScrollHandler allows to extend the selection by dragging the mouse
outside the display.  It does this by generating synthetic mouse move
events when the left mouse button is held down and the mouse is moved
outside of the display area.  This improves the user experience of
extending the selection to cover text that is scrolled outside the
display area.

Unfortunately, this was broken by commit
6667d96ed9d6c95eab4ed13449d4b134005ffb4b, which filters out mouse move
events if they don't change the character cell position.  So, modify the
filtering a bit, allowing mouse movements outside the display area to
pass through.