Create a test for block selection
This test currently is bogus, as soon as I reflow the Screen
object, the data is not correct. If anyone can check what I'm
doing wrong here it's appreciated.

The idea is that I pass a large string "aaa bbb ccc ddd" then
force a reflow for:

"aaa bbb
ccc ddd"

then I could copy only the column

bbb
ddd

Because the setSize did not work for that, the test is bogus.
but it's a start.