Add the SSH Manager Plugin
This plugin adds a SSH Manager, just like PuTTY, for Konsole.
It reads and parses the ~/.ssh/config file and generates entries
based on that, it also accepts entries manually.

The file used to store the information is a json, stored locally.
Clicking on the entry: request to connect to the server
Middle click on the entry: new tab, request to connect to the server

You can also filter entries, and disable / enable the view by
the new Plugins submenu.