Always keep "Show SSH Manager" text, let checkbox represent state
Having a checkable menu item combined with changing texts
is confusing. "Hide SSH Manager" in checked state would indicate
that the hidden state is active, which is not actually the case.

The menu item text should either indicate an action, in which case
renaming is okay but no checkbox should be added, or the
unchanging text should name the state that's either
active/checked or inactive/unchecked.

In the spirit of "Settings" -> "Show Menubar" / "Show Statusbar",
I'm picking the latter, so the text always remains
"Show SSH Manager" and only the checkbox state gets toggled.