When closing a session, don't close the whole window if there are splits
When closing a session, we check if that is the last tab, and make the code
close the whole window, but we also need to make sure it's the last view,
i.e. no split views.

CCBUG: 440976
FIXED-IN: 21.08.1
(cherry picked from commit 4a3cab03f5d853f4dd48531979fc3fb57dde5e2e)