Introduce a way to pass GUI focus to apps from the command line
Wayland has a mechanism (like X11 startupIDs) to get a token from a GUI
window with focus, then pass that to a new window. Which can use it to
raise itself.
On wayland we're more strict about making sure this is set.

Given we export environment variables about the view manager to the
shell, this exposes a method on that interface to return a token.

It can then be used with a simple alias such as:

```
XDG_ACTIVATION_TOKEN=`qdbus $KONSOLE_DBUS_SERVICE $KONSOLE_DBUS_WINDOW
org.kde.konsole.Window.activationToken` $@
```