Let the scrollbar follow the app color scheme
TerminalColor::onColorsChanged() used to be called updateScrollbarColors(),
now that KColorSchemeManager is integrated in Konsole, users can choose a
colour scheme for the whole window widget style, where tabs, scrollbar,
findbar, right-click context menu are all the same colour as you'd expect
from any KDE application.

This removes dimming of the scrollbar when the "dim when inactive" option is
disabled, the scrollbar is part of the window GUI, it should follow the same
color scheme.

2a71f063f8802a88171c87cfb63941317775d3d3
25e07b7e20df0827da6eb82dbe32adbe42aa4fdc
72d604c6cc96e0c2439244382ac07c6601970835