Refactor profile shortcuts handling code
- Don't use a QHash to hold a few profile shortcuts-related items, instead
use a vector of struct.

- Load profile shortcuts after loading the profiles; after commit
50041fd02ee42, all the profiles are loaded in the ProfileManager constructor.
Now there is no need to store the profile path as a member in ShortcutData;
and when saving the shortcut settings we can use profile->name(), since the
way the ProfileWriter works results in the filename being the
profile name + ".profile". When loading shortcuts, keep backwards
compatibility by checking for both "profile name" and "profile name.profile".