Remove pre-KDE4-era unimplemented Konsole-specific extension
CSI 28 ; <c> t is a pre-KDE4-era (from the comments) extension to
the control codes, where <c> was supposed to set the text-color
on a tab. It is unimplemented since KDE4, and only generated
a debug-message when used. Remove its implementation and the
TODO's associated with "document the meaning of this parameter".
Because there's no point in documenting an unimplemented thing.