Use smart pointers for color schemes.
Also uses weak pointers in the cache in ViewManager, to avoid having
them loaded for longer than necessary.

It takes less than a millisecond to load them, and it is nice to pick up
new changes from disk. But we don't want to do it for those currently
used by code elsewhere, so we still cache them with a weak_ptr that will
expire once all other code is done with it.