Store the current profile for session, and restore it later
For localhost, the user might have changed the current profile
manually, so before we trigger a profile change for a ssh session
we need to store that, so we can return to it later on.

If the user manually changes the profile inside of a ssh session.

TODO: test with multiple sessions (split screen, tabs)