ViewSplitter will now consistently create even-spaced splits
ViewSplitter was creating widgets of inconsistent sizes after adding/removing them in Konsole.
This change forces the active splitter to always be updated with the appropriate even-spacing size for the widgets.
As a drawback, manual changes to sizes will be overwritten when a new widget is added if no new splitter needs to be instantiated.

BUG: 447074