Let KColorSchemeManager handle loading/saving the widget color scheme
KColorSchemeManager (from KConfigWidgets) was changed in 5.89 to handle
loading/saving the selected color scheme automatically if
KColorSchemeManager::setAutosaveChanges(true) is used (it auto saves by
default, but calling setAutosaveChanges(true) makes it easier to understand
the code).

However this is conditional on KConfigWidgets >= 5.93 because that's where
the KColorSchemeManager::createSchemeSelectionMenu() overload used here was
changed to load the previously saved color scheme.

Since KColorSchemeManager uses "ColorScheme" as the config entry name,
migrate the old config entry when KF >= 5.93.