Support colon-separated sub-params in CSI and DCS
This adds supports for sub-parameters in control sequences. For now, it
supports the colon-separated variants of SGR 38 and SGR 48 for RGB and
indexed colours:

CSI 38 : 2 : <id> : <red> : <green> : <blue> [ : ... ] m
CSI 38 : 2 : <red> : <green> : <blue> m
CSI 38 : 5 : <index8> m

and likewise for SGR 48. <id> is ignored (it's supposed to be a colour
space id, to support different kinds of RGB).