URI: be more strict with www. URIs
We recognize URIs that start with an scheme and a possibly empty
authority, and URI suffixes that start with "www."

In the case of URIs starting with an scheme, they are of the form:

scheme://[ userinfo "@" ] host ...

while "www." URI suffixes are of the form:

www. <rest of host> ...

where host is actually in reg-name form (not in IPv4address or
IP-literal form).

This commit allows more strict parsing of e.g.

www.example.com:foo@bar.com

as <URI>:<email> instead of as a long <URI>.