Allow IPv6 literals in URIs
This allows recognizing URIs like the following:

http://[2a00:1450:4001:829::200e]/
http://[2a00:1450:4001:829::200e]:80/

Our regexp is not as strict as the syntax in RFC 3986, we just allow any
combination of hex-digits, colons and dots.

Besides, we don't go to the effort of forbidding IP-literals in www. URI
suffixes, so the following invalid suffix is recognized by our regexp:

www.[dead::beef]

but that was already recognized by the old regexp before 3b7e73f5.