Revert commit 60b0c9d
Commit 60b0c9d breaks bidi rendering, essentialy making all spaces
hard L2R characters, instead of neutral.

The deleted mostly does not do anything - it groups a
sequence of space which would be grouped anyway.

The problem is the break at the end, which exits the outer loop so
stops grouping after any sequence of space.

It "works" by having more groups, thus the rendering is closer to
rendering each character in its place, mitigating somewhat the issue
caused by non-fixed fonts (fallback to a font with different width).