Put tabs on top by default
Currently Konsole has its tabs on the bottom of the window. This is
nonstandard; few to no other main views in KDE apps have their tabs on
the bottom in the same manner.

In a productivity-focused app like Konsole, there could be an argument
for this being more Fitts' Law compliant, because tabs on the bottom
would touch the bottom screen edge when the window is maximized or
tiled and there is no bottom panel, making access via pointing device
fast. But... Plasma does have a bottom panel by default, so a person
would have to change their settings to realize this gain anyway. Such
a person could change their tabs to be on the bottom themselves.

There aren't strong reasons for putting the tabs on the bottom, so
let's move them to the top to match the behavior and appearance in
other KDE apps.