Fractional scaling and Antialias don't mix well
This is bug #373232 again.  The rendering rewrite in 76f879cd7 reenabled
antialasing in a too wide context, causing the bug to reappear.

After #373232, the situation was that QPainter::TextAntialiasing instead
of QPainter::Antialiasing would be enabled depending on the "Smooth
fonts" setting, and QPainter::Antialiasing would be enabled depending on
that same setting just for TerminalPainter::drawLineCharString().  Go
back to that situation.

BUG: 462914