Fix saving new profiles
When using the "Apply" button in the "Create new profile" dialog, the profile
was saved but the dialog remained in the "new profile" state. This meant that
when clicking on "Apply" again or accepting the dialog, it saved the profile
from scratch instead of editing the existing one. The ProfileManager
interpreted that as a rename because the "Name"/"UntranslatedName" properties
were set and decided to delete the profile at its "old" (== current) path.

BUG: 460383