Update blinking cursor in its correct visual position
The function to update the cursor when it blinks off has to update the
display according to visual cursor positions. Before 22.12, the cursor's
visual position was always identical to its logical, but now it may differ
when its inside an RTL string.

When painting the cursor, its visual position is known, so save it for
using in updateCursor().