SSHManager: Disable checking current process name in Windows
This allows to use ssh plugin in Windows by skipping the check
if the running process is shell or not. If the running process
is not a known shell, creating any new ssh connection will fail
without any error message. It is possible to run different shell
under cmd.exe or powershell.exe. So, it is not enough to check
only the child process of conpty's conhost.exe.