Fix a translucency repaint glitch in Wayland
When the konsole profile is set to use translucency and the konole part
is used in an app that doesn't support it, like konsole or dolphin, on X11
the translucency has no effect and the konsole part is opaque.
On wayland this doesn't happen and it tries to make that area of the app
translucent, but then the damage information is not correctly set and
there are drawing glitches in the transparent area.

This just disables transparency when the toplevel window doesn't have
the WA_TranslucentBackground attribute set.