Invalidate a slightly larger area when scheduling a repaint
In highdpi with fractional scaling factor some lines appear in the
terminal area. This may be the result of areas that should be
refreshed with new content that are not included in the invalidated
area. This commit increases the invalidated area a bit to ensure
the missing area is included.

CCBUG: 373232