Ensure both unerline and overline are drawn inside the row
This commit ensures both the underline and the overline are
both drawn inside the row. Each row keeps the same height as
the font.

If possible, the position of the underline and of the overline
are those provided by Qt. If the position is outside of the row,
then the nearest position is chosen.